No fanin inlining with good prex constraints
The current inliner will try to limit the inlining of methods with high
fanin. This heuristic is sound unless we have a FIXED_CLASS or
KNOWN_OBJECT constraint that improves on the signature of the callee
method. If we have improved type information we should continue trying
to inline due to the benefits that are likely to result due to the
additional context.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>