Flag calls that should be folded by VP
If the inliner find call where we want to abort inlining because the call
should be compile-time foldable by VP, we mark the node with a the flag
isFearGeneratingCall - this will be used to protect the call with OSR
guards. VP will omit the HCR guard for any transformations for calls
so marked.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>