Enhance the handling of indexOf in ValuePropagation
This change adds special case handling for the intrinsicIndexOfStringUTF16
JITHelper. The changes recognize when the target string is a constant length
of 0 or 1 and generates an appropriate inline sequence (iconst -1 or the
tree (icmpeq of the two chars) - 1). This avoids calling the helper for
a number of trivial examples present in java/net/URI$Parser.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>